# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SMART CLINIC - Healthcare SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments for clinic
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: doctorId
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /clinics/{clinicId}/appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update or reschedule appointment
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel appointment
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ===== ANALYTICS =====
...
  /clinics/{clinicId}/patients/{patientId}/emr:
    get:
      operationId: getPatientEmr
      tags: [patients]
      summary: Get full EMR timeline for a patient
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full EMR timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientEmr"

  # ===== BILLING PATHS =====
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ===== OTP / VERIFICATION SCHEMAS =====
...
    Prescription:
      type: object
      required: [id, clinicId, patientId, doctorId, prescriptionCode, medicationName, dosage, frequency, duration, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        doctorName:
          type: ["string", "null"]
        prescriptionCode:
          type: string
        medicationName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, dispensed, collected, cancelled]
        dispensedAt:
          type: ["string", "null"]
          format: date-time
        collectedAt:
          type: ["string", "null"]
          format: date-time
        dispensedById:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ===== QUEUE SCHEMAS =====
...
    NurseAssessment:
      type: object
      required: [id, clinicId, patientId, nurseId, triageLevel, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        nurseId:
          type: string
        nurseName:
          type: ["string", "null"]
        queueEntryId:
          type: ["string", "null"]
        bloodPressure:
          type: ["string", "null"]
        temperature:
          type: ["string", "null"]
        pulseRate:
          type: ["string", "null"]
        oxygenSaturation:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        bloodSugar:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        triageNotes:
          type: ["string", "null"]
        triageLevel:
          type: string
        createdAt:
          type: string
          format: date-time

    # ===== LAB SCHEMAS =====
...
    LabRequest:
      type: object
      required: [id, clinicId, patientId, doctorId, requestCode, testName, testCategory, urgency, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        patientName:
          type: ["string", "null"]
        patientCode:
          type: ["string", "null"]
        doctorId:
          type: string
        doctorName:
          type: ["string", "null"]
        requestCode:
          type: string
        testName:
          type: string
        testCategory:
          type: string
        urgency:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        createdAt:
          type: string
          format: date-time
...
    ConsultationNote:
      type: object
      required: [id, clinicId, patientId, doctorId, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        doctorName:
          type: ["string", "null"]
        consultationCode:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, pending_lab, pending_pharmacy, completed]
        queueEntryId:
          type: ["string", "null"]
        appointmentId:
          type: ["string", "null"]
        chiefComplaint:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        vitalSigns:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        prescription:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        followUpInstructions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ===== APPOINTMENT SCHEMAS =====
    AppointmentInput:
      type: object
      required: [patientId, doctorId, scheduledAt, type]
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        type:
          type: string
          enum: [consultation, follow_up, procedure, checkup, walk_in]
        visitReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        durationMinutes:
          type: integer
    AppointmentUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, confirmed, checked_in, waiting, in_consultation, completed, cancelled, no_show]
        visitReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        doctorId:
          type: string
    Appointment:
      type: object
      required: [id, clinicId, patientId, doctorId, scheduledAt, type, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        patientCode:
          type: string
        doctorId:
          type: string
        doctorName:
          type: string
        scheduledAt:
          type: string
          format: date-time
        checkedInAt:
          type: ["string", "null"]
          format: date-time
        type:
          type: string
        visitReason:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, confirmed, checked_in, waiting, in_consultation, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ===== ANALYTICS SCHEMAS =====
...
    Invoice:
      type: object
      required: [id, clinicId, patientId, invoiceCode, totalAmount, paidAmount, balance, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        patientCode:
          type: string
        doctorId:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        invoiceCode:
          type: string
        totalAmount:
          type: string
        paidAmount:
          type: string
        balance:
          type: string
        status:
          type: string
          enum: [unpaid, partial, paid, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PatientEmr:
      type: object
      required: [patientId, consultations, prescriptions, labRequests, appointments, assessments, invoices]
      properties:
        patientId:
          type: string
        consultations:
          type: array
          items:
            $ref: "#/components/schemas/ConsultationNote"
        prescriptions:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
        labRequests:
          type: array
          items:
            $ref: "#/components/schemas/LabRequest"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        assessments:
          type: array
          items:
            $ref: "#/components/schemas/NurseAssessment"
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"